自定义Security登录逻辑 合并 创建实例和保存实例内容接口 修改第三方登陆逻辑 使用uuid创建实例id

+++ springboot-dubbo-service/src/main/java/com/lzq/dubboservice/rabbitmq/QueueReceiver.java
@@ -10,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.amqp.rabbit.annotation.RabbitHandler;
 import org.springframework.amqp.rabbit.annotation.RabbitListener;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
 import java.io.File;
@@ -38,24 +39,30 @@ public class QueueReceiver {
     @Autowired
     AccountService accountService;
 
+    @Value("${resources.route}")
+    public String fileLocation;
+
     @RabbitHandler
     public void process(Example example){
         log.info(example.toString());
         SimpleDateFormat sdf = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
         log.info("消息接收时间:"+sdf.format(new Date()));
         //删除实例
-        exampleService.deleteExample(example.getExampleId());
-        File file = new File("F:/项目/MyDemo/data/" + example.getUsername() + "/" + example.getFileName()+".html");
-        //删除文件
-        file.delete();
-        //删除实例内容
-        contentService.deleteContent(example.getExampleId());
-        //删除实例图片（七牛云）
-        QiniuyunUtils.deleteFiles(example.getImg());
-        //删除所有用户的对该作品的喜爱
-        favoritesService.deleteFavorites(example.getExampleId());
-        //减少回收站数量（删除实例）
-        accountService.reduceRecycle(example.getUsername());
+        Boolean bol = exampleService.deleteExample(example.getExampleId());
+        if (bol){
+            File file = new File(fileLocation + example.getUsername() + "/" + example.getFileName()+".html");
+            //删除文件
+            file.delete();
+            //删除实例内容
+            contentService.deleteContent(example.getExampleId());
+            //删除实例图片（七牛云）
+            QiniuyunUtils.deleteFiles(example.getImg());
+            //删除所有用户的对该作品的喜爱
+            favoritesService.deleteFavorites(example.getExampleId());
+            //减少回收站数量（删除实例）
+            accountService.reduceRecycle(example.getUsername());
+        }
+
 
     }
